Make uv migration work with proper package structure
- Move all Python files into src/buzzinga/ package with __init__.py
- Convert all internal imports to relative imports
- Add [project.scripts] entry point so `uv run buzzinga` works
- Add hatchling build system with src layout config
- Move staticfiles/ into the package (bundled assets)
- Fix path resolution: BASE_PATH uses cwd() for data/, STATIC_FOLDER
  uses package dir for bundled assets, PyInstaller paths unchanged
- Recover missing user icon (deleted in 2021) as user.png
- Fix bogus os.path.join(ROOT_EXTENDED, STATIC_FOLDER) calls
- Add .venv/ and __pycache__/ to .gitignore
