+++ pyproject.toml
@@ -10,3 +10,13 @@ dependencies = [
     "pygame>=2.6.1",
     "pyinstaller>=6.16.0",
 ]
+
+[project.scripts]
+buzzinga = "buzzinga.buzzinga_class:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/buzzinga"]
