+++ src/buzzinga/whoknowsmore_quiz.py
@@ -1,9 +1,9 @@
 import os, random, pygame, json, sys, traceback
 
-from quiz_games import QuizGameBase
-from static import Static
-from game_utilities import blit_text_objects, optimize_text_in_container
-from animation import BuzzingaAnimation
+from .quiz_games import QuizGameBase
+from .static import Static
+from .game_utilities import blit_text_objects, optimize_text_in_container
+from .animation import BuzzingaAnimation
 
 
 class WhoKnowsMoreQuiz(QuizGameBase):
@@ -225,7 +225,7 @@ class WhoKnowsMoreQuiz(QuizGameBase):
                 pygame.display.flip()
                 if self.countdown_seconds_left == 0:
                     if self.is_game_sounds and not self.countdown_ended:
-                        countdown_sound = pygame.mixer.Sound(os.path.join(Static.ROOT_EXTENDED, Static.STATIC_FOLDER, 'countdown_end.wav'))
+                        countdown_sound = pygame.mixer.Sound(os.path.join(Static.STATIC_FOLDER, 'countdown_end.wav'))
                         self.game_sound_channel.play(countdown_sound)
                     self.draw_rect(Static.RED, Static.WHITE, 8, self.bottom_right_container)
                     pygame.display.flip()
